refactor: migrate to next experimental appDir structure
