+++ apps/scrum/next.config.js
@@ -6,6 +6,9 @@ const withBundleAnalyzer = require('@next/bundle-analyzer')({
 const config = {
   reactStrictMode: true,
   swcMinify: true,
+  experimental: {
+    appDir: true,
+  },
 }
 
 module.exports = withBundleAnalyzer(config)
