+++ apps/scrum/src/app/head.tsx
@@ -1,0 +1,8 @@
+export default function Head() {
+  return (
+    <>
+      <title>Scrum 新手村</title>
+      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    </>
+  )
+}
