+++ apps/scrum/src/app/layout.tsx
@@ -1,0 +1,11 @@
+import type { ChildrenProps } from 'react'
+
+import '../styles/globals.css'
+
+export default function RootLayout({ children }: ChildrenProps) {
+  return (
+    <html lang="zh-Hant-TW" className="h-full">
+      <body className="bg-white/90">{children}</body>
+    </html>
+  )
+}
