--- apps/scrum/src/pages/_app.tsx
@@ -1,7 +1,0 @@
-import '../styles/globals.css'
-
-import type { AppProps } from 'next/app'
-
-export default function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-}
