+++ apps/scrum/src/types/react.d.ts
@@ -1,0 +1,7 @@
+import React from 'react'
+
+declare module 'react' {
+  interface ChildrenProps {
+    children: React.ReactNode
+  }
+}
