+++ apps/scrum/test/index.spec.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react'
-import Home from '@/pages/index'
+import Home from '@/app/page'
 
 describe('home', () => {
   it('should have main', () => {
